Refactor RepairManagement
Closes #501
