+++ application/src/main/java/com/ericsson/bss/cassandra/ecchronos/application/spring/SpringBooter.java
@@ -15,7 +15,9 @@
 package com.ericsson.bss.cassandra.ecchronos.application.spring;
 
 import com.ericsson.bss.cassandra.ecchronos.rest.MetricsREST;
+import com.ericsson.bss.cassandra.ecchronos.rest.OnDemandRepairManagementRESTImpl;
 import com.ericsson.bss.cassandra.ecchronos.rest.RepairManagementRESTImpl;
+import com.ericsson.bss.cassandra.ecchronos.rest.ScheduleRepairManagementRESTImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.SpringApplication;
@@ -24,7 +26,8 @@ import org.springframework.boot.web.servlet.support.SpringBootServletInitializer
 import org.springframework.context.annotation.Import;
 
 @SpringBootApplication
-@Import(value = {RepairManagementRESTImpl.class, MetricsREST.class})
+@Import(value = {RepairManagementRESTImpl.class, ScheduleRepairManagementRESTImpl.class,
+        OnDemandRepairManagementRESTImpl.class, MetricsREST.class})
 public class SpringBooter extends SpringBootServletInitializer
 {
     private static final Logger LOG = LoggerFactory.getLogger(SpringBooter.class);
