+++ docs/autogenerated/openapi.yaml
@@ -9,10 +9,14 @@ servers:
 - url: https://localhost:8080
   description: Generated server url
 tags:
+- name: Repair-Management
+  description: View the status of repairs as well as run manual repairs
 - name: Metrics
   description: Retrieve metrics about ecChronos
 - name: Repair-Management
-  description: View the status of schedules and repairs as well as run manual repairs
+  description: Fetch repair information
+- name: Repair-Management
+  description: View the status of schedules
 - name: Actuator
   description: Monitor and interact
   externalDocs:
