+++ rest.osgi/src/main/java/com/ericsson/bss/cassandra/ecchronos/rest/osgi/RepairManagementRESTComponent.java
@@ -14,11 +14,7 @@
  */
 package com.ericsson.bss.cassandra.ecchronos.rest.osgi;
 
-import com.ericsson.bss.cassandra.ecchronos.core.repair.OnDemandRepairScheduler;
-import com.ericsson.bss.cassandra.ecchronos.core.repair.RepairScheduler;
-import com.ericsson.bss.cassandra.ecchronos.core.repair.types.OnDemandRepair;
 import com.ericsson.bss.cassandra.ecchronos.core.repair.types.RepairInfo;
-import com.ericsson.bss.cassandra.ecchronos.core.repair.types.Schedule;
 import com.ericsson.bss.cassandra.ecchronos.core.utils.RepairStatsProvider;
 import com.ericsson.bss.cassandra.ecchronos.core.utils.ReplicatedTableProvider;
 import com.ericsson.bss.cassandra.ecchronos.core.utils.TableReferenceFactory;
@@ -32,7 +28,6 @@ import org.osgi.service.component.annotations.ReferencePolicy;
 import org.springframework.http.ResponseEntity;
 
 import java.time.Duration;
-import java.util.List;
 
 /**
  * OSGi component wrapping {@link RepairManagementREST} bound with OSGi services.
@@ -40,16 +35,6 @@ import java.util.List;
 @Component
 public class RepairManagementRESTComponent implements RepairManagementREST
 {
-    @Reference (service = RepairScheduler.class,
-            cardinality = ReferenceCardinality.MANDATORY,
-            policy = ReferencePolicy.STATIC)
-    private volatile RepairScheduler myRepairScheduler;
-
-    @Reference (service = OnDemandRepairScheduler.class,
-            cardinality = ReferenceCardinality.MANDATORY,
-            policy = ReferencePolicy.STATIC)
-    private volatile OnDemandRepairScheduler myOnDemandRepairScheduler;
-
     @Reference(service = TableReferenceFactory.class,
             cardinality = ReferenceCardinality.MANDATORY,
             policy = ReferencePolicy.STATIC)
@@ -69,42 +54,8 @@ public class RepairManagementRESTComponent implements RepairManagementREST
     @Activate
     public final synchronized void activate()
     {
-        myDelegateRESTImpl = new RepairManagementRESTImpl(myRepairScheduler, myOnDemandRepairScheduler,
-                myTableReferenceFactory, myReplicatedTableProvider, myRepairStatsProvider);
-    }
-
-    @Override
-    public final ResponseEntity<List<OnDemandRepair>> getRepairs(final String keyspace,
-                                                                 final String table,
-                                                                 final String hostId)
-    {
-        return myDelegateRESTImpl.getRepairs(keyspace, table, hostId);
-    }
-
-    @Override
-    public final ResponseEntity<List<OnDemandRepair>> getRepairs(final String id, final String hostId)
-    {
-        return myDelegateRESTImpl.getRepairs(id, hostId);
-    }
-
-    @Override
-    public final ResponseEntity<List<Schedule>> getSchedules(final String keyspace, final String table)
-    {
-        return myDelegateRESTImpl.getSchedules(keyspace, table);
-    }
-
-    @Override
-    public final ResponseEntity<Schedule> getSchedules(final String id, final boolean full)
-    {
-        return myDelegateRESTImpl.getSchedules(id, full);
-    }
-
-    @Override
-    public final ResponseEntity<List<OnDemandRepair>> triggerRepair(final String keyspace,
-                                                                    final String table,
-                                                                    final boolean isLocal)
-    {
-        return myDelegateRESTImpl.triggerRepair(keyspace, table, isLocal);
+        myDelegateRESTImpl = new RepairManagementRESTImpl(myTableReferenceFactory,
+                myReplicatedTableProvider, myRepairStatsProvider);
     }
 
     @Override
